Actulizacion de getsetvarside en programas turisticos
